Create README.md
The original data (lashelterdata.rda) comes from shelters in the LA area and spans from mid 2009 to mid 2013. In this investigation, I used the data to identify trends that dog owners looking to surrender their pets can use to increase the likelihood of adoption for their dogs. The second file, report.rmd, is the R markdown file which contains the code. The third file, report.html, is a readable report detailing the finding of the investigation.